Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int a,b,product;
-    scanf("%.2f %.2f",&a, &b);
+    float a,b,product;
+    scanf("%f %f",&a, &b);
     product=a*b;
-    printf("Product: %f",product);
+    printf("Product: %.2f",product);
     return 0;
 }
